requirements.txt
django==4.2.5
djangorestframework==3.14.0
psycopg2-binary==2.9.9

